Move Product events to application layer
